fix: add more styles
