+++ nextjs-blog/pages/global.css
@@ -22,3 +22,27 @@
     border-radius: 3px;
     padding: 1px 3px;
 }
+
+.content h2 {
+    font-size: 1.5rem;
+    border-bottom: 1px solid;
+    margin-bottom: 1rem;
+}
+
+.content h3 {
+    font-size: 1.3rem;
+    border-bottom: 1px solid;
+    margin-bottom: 1rem;
+}
+
+.content h4 {
+    font-size: 1.1rem;
+    border-bottom: 1px solid;
+    margin-bottom: 1rem;
+}
+
+.content h5,h6 {
+    font-size: 1rem;
+    border-bottom: 1px solid;
+    margin-bottom: 1rem;
+}
