+++ nextjs-blog/pages/post/[year]/[month]/[day]/[title]/index.tsx
@@ -1,34 +1,11 @@
-import matter from "gray-matter"
 import path from "path"
 import fs from "fs"
-import remark from "remark"
-import html from "remark-html"
 import Head from "next/head"
 
 import { Header } from "../../../../../../components/Header"
+import { getPostData } from "../../../../../../lib/markdown"
 
 const postsDirectory = path.join(process.cwd(), '../source/_posts/')
-
-export async function getPostData(params: { title: string, year: string, month: string, day: string }) {
-    const fullPath = path.join(postsDirectory, `${params.year}-${params.month}-${params.day}-${params.title}.markdown`)
-    const fileContents = fs.readFileSync(fullPath, 'utf8')
-  
-    // Use gray-matter to parse the post metadata section
-    const matterResult = matter(fileContents)
-  
-    // Use remark to convert markdown into HTML string
-    const processedContent = await remark()
-      .use(html)
-      .use(require('remark-linkify-regex')(/https?:\/\/[^\s]*/))
-      .process(matterResult.content)
-    const contentHtml = processedContent.toString()
-  
-    // Combine the data with the id and contentHtml
-    return {
-      contentHtml,
-      ...matterResult.data
-    }
-}
 
 export async function getStaticPaths() {
     const fileNames = fs.readdirSync(postsDirectory)
@@ -43,7 +20,7 @@ export async function getStaticPaths() {
 }
 
 export async function getStaticProps({ params }: any) {
-    const postData = await getPostData(params)
+    const postData = await getPostData(params);
     return {
         props: {
             postData
@@ -62,7 +39,7 @@ export default function Post({ postData }: any) {
         <article className="container mx-auto max-w-5xl p-4">
             <h1 className="text-3xl py-4">{postData.title}</h1>
             <p>{postData.date}</p>
-            <div className="content" dangerouslySetInnerHTML={{ __html: postData.contentHtml}}></div>
+            <div className="content py-8" dangerouslySetInnerHTML={{ __html: postData.contentHtml}}></div>
         </article>
         </>
     )
