+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "remark-prism": "^1.3.6"
+  }
+}
